requirements.txt
# 核心依赖
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0

# 音频处理（如果需要）
librosa>=0.10.0
soundfile>=0.12.0

# 视频处理（如果需要）
opencv-python>=4.8.0
imageio>=2.31.0

# 评估指标
scikit-learn>=1.3.0

# 工具
tqdm>=4.66.0
pyyaml>=6.0

# Hugging Face数据集
datasets>=2.14.0

